requirements-debug.txt
-r base.txt

pytest-django
pytest-xdist
pytest-cov